Final_proj/requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
python-dotenv
Flask-JWT-Extended
transformers
torch
nltk
sentence-transformers